+++ app/controllers/api/v1/munchies_controller.rb
@@ -4,11 +4,11 @@ class Api::V1::MunchiesController < Api::V1::ApiBaseController
     # require 'pry'; binding.pry
 
     duration = DurationService.new.get_duration(params[:start], params[:end])
-
-    restaurants = RestaurantService.new.get_restaurants(params[:start], params[:end], duration)
-
-
     require 'pry'; binding.pry
+
+    restaurants = RestaurantService.new.get_restaurants(params, duration)
+
+
 
     # render json: MunchiesSerializer.new(unfiltered_forecast,
                                                # location[:citystate],
