+++ spec/services/restaurant_service_spec.rb
@@ -4,7 +4,23 @@ describe RestaurantService do
   it 'returns destination and three open restaurants at ETA' do
     service = RestaurantService.new
 
-    duration = service.get_restaurants('denver,co', 'pueblo,co', )
+    duration = {
+      text: "1 hour 47 mins",
+      value: 6413
+    }
+
+    restaurant_params = {
+      start: "denver,co",
+      end: "pueblo,co",
+      food: "chinese"
+    }
+    # restaurant_params = {
+    #   "start"=>"denver,co",
+    #   "end"=>"pueblo,co",
+    #   "food"=>"chinese"
+    # }
+
+    duration = service.get_restaurants(restaurant_params, duration)
 
     expect(duration.class).to eq(Hash)
     expect(duration).to have_key(:text)
